reorganized class structure
